+++ src/app/(route)/main/_components/liveItem/NewPostItem.tsx
