+++ src/app/(route)/main/_components/news/newsItem.tsx
