+++ src/app/(route)/main/_components/state/EmptyGameBox.tsx
