+++ src/app/(route)/main/_components/state/EmptyScheduleItem.tsx
