+++ src/app/(route)/main/_components/state/EventItem.tsx
