+++ src/app/(route)/main/_components/state/EventItemSkeleton.tsx
