+++ src/app/(route)/main/_components/state/MainBigSizeNewsSkeleton.tsx
