+++ src/app/(route)/main/_components/state/MainPostItemSkeleton.tsx
