+++ src/app/(route)/main/_components/state/NewsItemSkeleton.tsx
